- findNumberOfEqualIndexes methods fixed - randi methods fixed according to the matlab function

+++ DuckShootingTest.hh
@@ -59,8 +59,8 @@ public:
     DuckShootingTest();
     ~DuckShootingTest();
     bool PoissonRandom(double);
-    std::vector<std::vector<double>> randi(uint32_t numberRange, uint32_t numOfRows, uint32_t numOfColmns);
-    uint32_t findNumberOfEqualIndexes(std::vector<std::vector<double>> &targets, uint32_t nn);
+    std::vector<std::vector<int>> randi(uint32_t numberRange, uint32_t numOfRows, uint32_t numOfColmns);
+    uint32_t findNumberOfEqualIndexes(std::vector<std::vector<int>> &targets, uint32_t nn);
     std::vector<std::vector<double>> DuckShootingTest::MatrixFillWith(uint32_t numOfRows, uint32_t numOfColmns, double filledWith);
 };
